requirements.txt
Django>=1.11, <2.0
djangorestframework==3.6.4
django-debug-toolbar
scikit-image
django-allauth
django-suit==0.2.25